Refactor user directory changes
